end class of friday create animation on cart_view creare the result_page more than adjusts in the style class

+++ lib/shared/models/user_model.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 class UserModel {
   final String name;
   final String photoUrl;
-  final int score;
+  final double score;
 
   UserModel({
     required this.name,
